+++ ethereum/rpc/namespaces/net/api.go
@@ -1,16 +1,18 @@
 package net
 
 import (
+	"context"
 	"fmt"
 
-	ethermint "github.com/tharsis/ethermint/types"
-
 	"github.com/cosmos/cosmos-sdk/client"
+	rpcclient "github.com/tendermint/tendermint/rpc/client"
+	ethermint "github.com/tharsis/ethermint/types"
 )
 
 // PublicAPI is the eth_ prefixed set of APIs in the Web3 JSON-RPC spec.
 type PublicAPI struct {
 	networkVersion uint64
+	tmClient       rpcclient.Client
 }
 
 // NewPublicAPI creates an instance of the public Net Web3 API.
@@ -23,10 +25,31 @@ func NewPublicAPI(clientCtx client.Context) *PublicAPI {
 
 	return &PublicAPI{
 		networkVersion: chainIDEpoch.Uint64(),
+		tmClient:       clientCtx.Client,
 	}
 }
 
 // Version returns the current ethereum protocol version.
 func (s *PublicAPI) Version() string {
 	return fmt.Sprintf("%d", s.networkVersion)
+}
+
+// Listening returns if client is actively listening for network connections.
+func (s *PublicAPI) Listening() bool {
+	ctx := context.Background()
+	netInfo, err := s.tmClient.NetInfo(ctx)
+	if err != nil {
+		return false
+	}
+	return netInfo.Listening
+}
+
+// PeerCount returns the number of peers currently connected to the client.
+func (s *PublicAPI) PeerCount() int {
+	ctx := context.Background()
+	netInfo, err := s.tmClient.NetInfo(ctx)
+	if err != nil {
+		return 0
+	}
+	return len(netInfo.Peers)
 }